Simplify start debug command handler
